started migrating app/Controllers logic to api/Controllers

+++ api/Controllers/V1/SettingController.php
@@ -5,6 +5,7 @@ namespace Api\Controllers\V1;
 use Api\Controllers\BaseController;
 use Lib\Models\Setting;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 
 class SettingController extends BaseController
 {
@@ -26,42 +27,48 @@ class SettingController extends BaseController
      */
     public function store(Request $request)
     {
-        return Setting::create($request->except(["token","_token"]));
+      if(!$this->user()->hasPermissionTo("create settings"))
+        throw new UnauthorizedHttpException("you can't update a setting. Please contact administrator");
+      return Setting::create($request->except(["token","_token"]));
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Settings  $settings
-     * @return \Illuminate\Http\Response
-     */
+  
+  /**
+   * Display the specified resource.
+   *
+   * @param Setting $setting
+   * @return \Illuminate\Http\Response|Setting
+   */
     public function show(Setting $setting)
     {
         return $setting;
     }
-
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Settings  $settings
-     * @return \Illuminate\Http\Response
-     */
+  
+  
+  /**
+   * Update the specified resource in storage.
+   *
+   * @param  \Illuminate\Http\Request $request
+   * @param Setting $settings
+   * @return \Illuminate\Http\Response|Setting
+   */
     public function update(Request $request, Setting $settings)
     {
+        if(!$this->user()->hasPermissionTo("edit settings"))
+          throw new UnauthorizedHttpException("you can't update a setting. Please contact administrator");
         $settings->update($request->except(["token","_token"]));
         return $settings;
     }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Settings  $settings
-     * @return \Illuminate\Http\Response
-     */
+  
+  /**
+   * Remove the specified resource from storage.
+   *
+   * @param Setting $settings
+   * @return \Illuminate\Http\Response
+   */
     public function destroy(Setting $settings)
     {
-        $settings->delete();
+      if(!$this->user()->hasPermissionTo("delete settings"))
+        throw new UnauthorizedHttpException("you can't update a setting. Please contact administrator");
+      $settings->delete();
     }
 }
